Finished Menu Template Scene
Improved the template I brought in and designed it from easier game growth and reuse
